Agregado middleware auth a los controladores, excepto al controlador de inicio para el cual no es necesario estar autentificado
