+++ app/Http/Controllers/DoctorsController.php
@@ -6,6 +6,11 @@ use Illuminate\Http\Request;
 
 class DoctorsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
         //
